# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Just-In-Time Consultancy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export type ServiceRequestStatus = typeof ServiceRequestStatus[keyof typeof ServiceRequestStatus];


export const ServiceRequestStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface ServiceRequest {
  id: number;
  requestId: string;
  companyName: string;
  contactPerson: string;
  email: string;
  phone: string;
  country: string;
  productName: string;
  productCategory: string;
  description: string;
  quantity: string;
  expectedBudget: string;
  preferredDeliveryCountry: string;
  /** @nullable */
  requiredDeliveryDate?: string | null;
  /** @nullable */
  additionalNotes?: string | null;
  /** @nullable */
  attachmentFileName?: string | null;
  /** @nullable */
  attachmentData?: string | null;
  status: ServiceRequestStatus;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ServiceRequestInput {
  companyName: string;
  contactPerson: string;
  email: string;
  phone: string;
  country: string;
  productName: string;
  productCategory: string;
  description: string;
  quantity: string;
  /** @nullable */
  expectedBudget?: string | null;
  preferredDeliveryCountry: string;
  /** @nullable */
  requiredDeliveryDate?: string | null;
  /** @nullable */
  additionalNotes?: string | null;
  /** @nullable */
  attachmentFileName?: string | null;
  /** @nullable */
  attachmentData?: string | null;
}

export type ServiceRequestUpdateStatus = typeof ServiceRequestUpdateStatus[keyof typeof ServiceRequestUpdateStatus];


export const ServiceRequestUpdateStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface ServiceRequestUpdate {
  status?: ServiceRequestUpdateStatus;
  /** @nullable */
  adminNotes?: string | null;
}

export interface ServiceRequestList {
  requests: ServiceRequest[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
}

export interface AdminAuthResult {
  message: string;
  admin: AdminUser;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  createdAt: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface MonthlyCount {
  month: string;
  count: number;
}

export interface AnalyticsData {
  totalRequests: number;
  pendingRequests: number;
  inProgressRequests: number;
  completedRequests: number;
  cancelledRequests: number;
  totalContactMessages: number;
  totalSupplierRegistrations: number;
  recentRequests: ServiceRequest[];
  recentContactMessages: ContactMessage[];
  requestsByCategory: CategoryCount[];
  requestsByMonth: MonthlyCount[];
}

export interface ContactInput {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
}

export interface ContactMessageList {
  messages: ContactMessage[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface ExportResponse {
  data: ServiceRequest[];
  filename: string;
}

export type ListAdminRequestsParams = {
status?: ListAdminRequestsStatus;
search?: string;
page?: number;
limit?: number;
};

export type ListAdminRequestsStatus = typeof ListAdminRequestsStatus[keyof typeof ListAdminRequestsStatus];


export const ListAdminRequestsStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type ExportAdminRequestsParams = {
status?: ExportAdminRequestsStatus;
};

export type ExportAdminRequestsStatus = typeof ExportAdminRequestsStatus[keyof typeof ExportAdminRequestsStatus];


export const ExportAdminRequestsStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type ListAdminMessagesParams = {
search?: string;
page?: number;
limit?: number;
};