# lib/api-zod/src/generated/types/analyticsData.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Just-In-Time Consultancy API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from './categoryCount';
import type { ContactMessage } from './contactMessage';
import type { MonthlyCount } from './monthlyCount';
import type { ServiceRequest } from './serviceRequest';

export interface AnalyticsData {
  totalRequests: number;
  pendingRequests: number;
  inProgressRequests: number;
  completedRequests: number;
  cancelledRequests: number;
  totalContactMessages: number;
  totalSupplierRegistrations: number;
  recentRequests: ServiceRequest[];
  recentContactMessages: ContactMessage[];
  requestsByCategory: CategoryCount[];
  requestsByMonth: MonthlyCount[];
}